+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ### Added
 
 - Support for `EXEC_BACKEND` builds ([PG-2547](https://perconadev.atlassian.net/browse/PG-2547))
+
+
+### Changed
+
+- Do not leave the `pgsm_create_view()` function after running `CREATE EXTENSION`